certificates/requirements.txt
google-cloud-firestore==2.1.3
firebase-admin==4.5.3
google-cloud-core==1.6.0
pillow